pipeline/requirements.txt
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
psycopg2-binary==2.9.9
python-decouple==3.8
schedule==1.2.0
sqlalchemy==2.0.23
python-dateutil==2.8.2